Matern Generalized Wendland (GW) covariance function with kappa = 2 (scalar - generic)

adapted from Bevilacqua et al. 2019.

Usage

single_gw2(d, sigsq, phi, mu)

Arguments

d

a scalar representing the distance on which it is desired to evaluate the covariance function.

sigsq

the \(\sigma^2\) parameter from the Matern covariance function.

phi

the \(\phi\) parameter from the Matern covariance function, controls the range of the spatial dependence.

mu

a parameter that controls the smoothness of the covariance function. Note that, \(\mu \geq 1\).

Value

a scalar representing the GW covariance between two observations d apart of each other.