Spherical covariance function (scalar)

Computing the Spherical covariance function for a single distance measure.

Usage

single_spher(d, sigsq, phi)

Arguments

d

a scalar representing the distance on which it is desired to evaluate the covariance function.

sigsq

the \(\sigma^2\) parameter from the Spherical covariance. function.

phi

the \(\phi\) parameter from the Spherical covariance function, controls the range of the spatial dependence.

Value

a scalar representing the (gaussian) covariance between two observations d apart of each other.

See also