SMILE main function

This function encompasses the original functions and assumptions. It outputs a time series for each of the compartments given a set of parameters.

Usage

smile_main(
  b0,
  b1,
  period,
  theta,
  tau,
  years,
  alpha = 1/52,
  zeta_novax = 0.88,
  gamma = 0.9868,
  sigmaa = 1,
  psi = 1,
  N1 = 5000,
  K = 5000,
  vax = NULL,
  beta_0 = NULL,
  beta_1 = NULL,
  b_fixed = NULL,
  age_struc = TRUE,
  stochastic = FALSE,
  LIZ_init = 1,
  rho_pop = NULL,
  output_df = FALSE
)

Arguments

b0

strength of seasonality parameter

b1

strength of seasonality parameter

period

periodicity of the infection in weeks

theta

dispersion effort parameter

tau

dispersion effort parameter

years

number of years to run the simulation

alpha

probability of transitioning from immune to susceptible in one week timestep

zeta_novax

probability of becoming immune after infection (fixed when no vaccination)

gamma

spore persistence rate

sigmaa

rate of death by other causes than disease

psi

number of spores per carcass

N1

starting population size

K

carrying capacity for population dynamics

vax

vaccination rates for each year

beta_0

survival parameter

beta_1

vaccine survival parameter

b_fixed

Number of infections caused by one LIZ when dispersion effort is zero and assume no seasonality

age_struc

default is TRUE, if set to false, no population dynamics are considered

stochastic

default is FALSE, so the simulations are deterministic

LIZ_init

Initial number of carcasses, set to default 1

rho_pop

average reproduction rate to be used when simulating with population dynamics

output_df

if set to TRUE, will return a dataframe instead of a list

Value

description

Examples

head(SMILE::smile_main(-30, 0.85, 3*52, 10, 10, 10, output_df = TRUE))
#>   week    S M I L        E
#> 1    1 5000 0 0 1 1.000000
#> 2    2 5000 0 0 0 1.986800
#> 3    3 5000 0 0 0 1.960574
#> 4    4 5000 0 0 0 1.934695
#> 5    5 5000 0 0 0 1.909157
#> 6    6 5000 0 0 0 1.883956