Generalized Wendland covariance function for a given distance matrix.

Computing the Matern covariance function for a matrix of distances.

Usage

gw_cov(dists, sigsq, phi, kappa, mu)

Arguments

dists

a numeric matrix representing the distance between spatial entities.

sigsq

the \(\sigma^2\) parameter from the Matern covariance function.

phi

\(\phi\) is the range of the covariance function.

kappa

\(\kappa \in \{0, \ldots, 3 \}\).

mu

\(\mu\) controls the smoothness of the covariance function

Value

The GW (isotropic) covariance function associated with the provided distances (dists) and the given set of parameters.