sf to spm

Transforming one (or two) sf objects into a sspm (or mspm) object

sf_to_spm(
  sf_obj1,
  sf_obj2,
  n_pts,
  type = "regular",
  by_polygon = FALSE,
  poly_ids = NULL,
  var_ids
)

Arguments

sf_obj1
a sf object s.t. its geometries are polygons.

sf_obj2
a sf object s.t. its geometries are polygons.

n_pts
a numeric scalar representing the number of points to create a grid in the study region on which the polygons in sf_obj is observed. Alternatively, it can be a vector of the same length as nrow(sf_obj). In this case, it generates the given number of points for each polygon in sf_obj. Considering the scenario on which you also input a sf_obj2, this argument can be set to be a list with two positions such that the first position contains the n_pts argument for sf_obj1, and the second position contains informations about the same parameter but for the sf_obj2.

type
a character indicating the type of grid to be generated. The options are c("random", "regular", "hexagonal"). For more details, see st_sample. Similarly to n_pts, you can input a two positions list in here too.

by_polygon
a logical indicating wheter we should generate n_pts by polygon or for the n_pts for the whole study region. Similarly to n_pts (and by_polygon), you can input a two positions list in here too.

poly_ids
a list with two positions, being the first and the second one a character vector informing the name of the variable in sf_obj and sf_obj2, respectively, that represents the polygons unique identifiers. In case this is not informed, we assume the id of the polygons are given by their row numbers. Note that, if this is informed, it must be a two positions list.

var_ids
a two positions list containing in each position either a scalar or vector of type character, indicating the (numerical) variables that are going to be analyzed for sf_obj1 and sf_obj2, respectively.

Value

a sspm or mspm object.